application/kafka-producer/requirements.txt
kafka-python==2.0.2
schedule==1.1.0
aiokafka==0.7.2
Faker==15.1.3
requests==2.32.3
pyyaml==6.0.1
python-dotenv==1.0.0